Refactor RX47 combat mech systems
