+++ Content.Shared/_Stories/CombatMech/CombatMechSystem.Pilot.cs
@@ -1,0 +1,233 @@
+using System.Numerics;
+using Content.Shared._RMC14.Hands;
+using Content.Shared._RMC14.Attachable.Components;
+using Content.Shared._RMC14.Attachable.Events;
+using Content.Shared._RMC14.Atmos;
+using Content.Shared._RMC14.BlurredVision;
+using Content.Shared._RMC14.CameraShake;
+using Content.Shared._RMC14.Damage;
+using Content.Shared._RMC14.Entrenching;
+using Content.Shared._RMC14.Explosion;
+using Content.Shared._RMC14.Marines;
+using Content.Shared._RMC14.Marines.Skills;
+using Content.Shared._RMC14.Slow;
+using Content.Shared._RMC14.Stealth;
+using Content.Shared._RMC14.Stun;
+using Content.Shared._RMC14.Weapons.Ranged;
+using Content.Shared._RMC14.Weapons.Ranged.IFF;
+using Content.Shared._RMC14.Xenonids.Acid;
+using Content.Shared._RMC14.Xenonids.Neurotoxin;
+using Content.Shared._RMC14.Xenonids.Paralyzing;
+using Content.Shared._RMC14.Xenonids.Parasite;
+using Content.Shared._RMC14.Xenonids.Weeds;
+using Content.Shared.Atmos.Components;
+using Content.Shared.Buckle;
+using Content.Shared.Buckle.Components;
+using Content.Shared.Containers;
+using Content.Shared.Containers.ItemSlots;
+using Content.Shared.Damage;
+using Content.Shared.DoAfter;
+using Content.Shared.Examine;
+using Content.Shared.Explosion;
+using Content.Shared.FixedPoint;
+using Content.Shared.Hands;
+using Content.Shared.Hands.Components;
+using Content.Shared.Hands.EntitySystems;
+using Content.Shared.Interaction;
+using Content.Shared.Interaction.Components;
+using Content.Shared.Interaction.Events;
+using Content.Shared.Item;
+using Content.Shared.Movement.Components;
+using Content.Shared.Movement.Events;
+using Content.Shared.Movement.Systems;
+using Content.Shared.Mobs.Systems;
+using Content.Shared.Physics;
+using Content.Shared.Popups;
+using Content.Shared.Standing;
+using Content.Shared.StatusEffect;
+using Content.Shared.StatusEffectNew;
+using Content.Shared.Stunnable;
+using Content.Shared.Verbs;
+using Content.Shared.Weapons.Ranged.Systems;
+using Robust.Shared.Audio.Systems;
+using Robust.Shared.Containers;
+using Robust.Shared.Map;
+using Robust.Shared.Network;
+using Robust.Shared.Physics;
+using Robust.Shared.Physics.Events;
+using Robust.Shared.Physics.Systems;
+using Robust.Shared.Timing;
+
+namespace Content.Shared._Stories.CombatMech;
+
+
+public sealed partial class CombatMechSystem
+{
+    private void OnStrapAttempt(Entity<CombatMechComponent> ent, ref StrapAttemptEvent args)
+    {
+        if (args.Cancelled)
+            return;
+
+        if (args.User is not { } user)
+            return;
+
+        if (!_skills.HasSkill(user, ent.Comp.WeaponSkill, ent.Comp.WeaponSkillRequired))
+        {
+            if (args.Popup)
+                _popup.PopupClient(Loc.GetString("stories-rx47-not-trained-pilot"), ent, user, PopupType.MediumCaution);
+            args.Cancelled = true;
+            return;
+        }
+
+        if (GetWeapon(ent, true) == null || GetWeapon(ent, false) == null)
+        {
+            if (args.Popup)
+                _popup.PopupClient(Loc.GetString("stories-rx47-missing-weapons"), ent, user, PopupType.MediumCaution);
+            args.Cancelled = true;
+            return;
+        }
+
+        if (!TryComp(user, out HandsComponent? hands) || _hands.CountFreeHands((user, hands)) < 2)
+        {
+            if (args.Popup)
+                _popup.PopupClient(Loc.GetString("stories-rx47-need-both-hands"), ent, user, PopupType.MediumCaution);
+            args.Cancelled = true;
+        }
+    }
+
+    private void OnUnstrapAttempt(Entity<CombatMechComponent> ent, ref UnstrapAttemptEvent args)
+    {
+        if (args.Cancelled || !ent.Comp.HelmetClosed)
+            return;
+
+        if (args.Popup)
+            _popup.PopupClient(Loc.GetString("stories-rx47-faceplate-blocks-exit"), ent, args.User, PopupType.MediumCaution);
+
+        args.Cancelled = true;
+    }
+
+    private void OnStrapped(Entity<CombatMechComponent> ent, ref StrappedEvent args)
+    {
+        var pilot = args.Buckle.Owner;
+
+        var inside = EnsureComp<InsideCombatVehicleComponent>(pilot);
+        inside.Vehicle = ent;
+        Dirty(pilot, inside);
+        UpdatePilotProtection((pilot, inside));
+
+        _mover.SetRelay(pilot, ent);
+        var relay = EnsureComp<InteractionRelayComponent>(pilot);
+        _interaction.SetRelay(pilot, ent, relay);
+
+        _audio.PlayPredicted(ent.Comp.EnterSound, ent, pilot);
+        _movementSpeed.RefreshMovementSpeedModifiers(ent);
+
+        if (_net.IsServer)
+        {
+            if (!TransferWeaponToPilot(ent, pilot, true) ||
+                !TransferWeaponToPilot(ent, pilot, false))
+            {
+                EjectPilotAfterWeaponTransferFailure(ent, pilot);
+                return;
+            }
+        }
+
+        UpdateAppearance(ent);
+    }
+
+    private void OnUnstrapped(Entity<CombatMechComponent> ent, ref UnstrappedEvent args)
+    {
+        var pilot = args.Buckle.Owner;
+
+        if (TryComp(pilot, out InsideCombatVehicleComponent? inside))
+            RestorePilotProtection((pilot, inside));
+
+        RemCompDeferred<InsideCombatVehicleComponent>(pilot);
+        RemComp<RelayInputMoverComponent>(pilot);
+        RemComp<MovementRelayTargetComponent>(ent);
+        RemCompDeferred<InteractionRelayComponent>(pilot);
+        _movementSpeed.RefreshMovementSpeedModifiers(ent);
+
+        if (_net.IsServer)
+        {
+            TransferWeaponToMech(ent, pilot, true);
+            TransferWeaponToMech(ent, pilot, false);
+        }
+
+        _audio.PlayPredicted(ent.Comp.EnterSound, ent, pilot);
+        UpdateAppearance(ent);
+    }
+
+    private bool TransferWeaponToPilot(Entity<CombatMechComponent> ent, EntityUid pilot, bool primary)
+    {
+        if (GetWeapon(ent, primary) is not { } weapon)
+            return false;
+
+        if (!TryComp(pilot, out HandsComponent? pilotHands))
+            return false;
+
+        var hand = FindHand(pilot, pilotHands, primary ? HandLocation.Left : HandLocation.Right);
+        if (hand == null)
+            return false;
+
+        LinkWeaponToMech(weapon, ent);
+
+        RemComp<UnremoveableComponent>(weapon);
+        if (_hands.IsHolding(ent.Owner, weapon))
+            _hands.TryDrop(ent.Owner, weapon, Transform(ent).Coordinates, checkActionBlocker: false, doDropInteraction: false);
+
+        if (!_hands.TryPickup(pilot, weapon, hand, checkActionBlocker: false, animate: false, handsComp: pilotHands))
+        {
+            TransferWeaponToMech(ent, pilot, primary);
+            return false;
+        }
+
+        EnsureComp<UnremoveableComponent>(weapon);
+        return true;
+    }
+
+    private void TransferWeaponToMech(Entity<CombatMechComponent> ent, EntityUid pilot, bool primary)
+    {
+        if (GetWeapon(ent, primary) is not { } weapon)
+            return;
+
+        RemComp<UnremoveableComponent>(weapon);
+
+        if (!TryComp(ent.Owner, out HandsComponent? mechHands))
+            return;
+
+        var hand = FindHand(ent.Owner, mechHands, primary ? HandLocation.Left : HandLocation.Right);
+        if (hand == null)
+            return;
+
+        LinkWeaponToMech(weapon, ent);
+
+        if (_hands.IsHolding(ent.Owner, weapon))
+        {
+            EnsureComp<UnremoveableComponent>(weapon);
+            return;
+        }
+
+        if (_hands.IsHolding(pilot, weapon))
+            _hands.TryDrop(pilot, weapon, Transform(ent).Coordinates, checkActionBlocker: false, doDropInteraction: false);
+
+        if (!_hands.TryPickup(ent.Owner, weapon, hand, checkActionBlocker: false, animate: false, handsComp: mechHands))
+        {
+            _transform.SetCoordinates(weapon, Transform(ent).Coordinates);
+            Log.Warning($"RX47 failed to return {ToPrettyString(weapon)} to {ToPrettyString(ent.Owner)} hand {hand}.");
+            return;
+        }
+
+        EnsureComp<UnremoveableComponent>(weapon);
+    }
+
+    private void EjectPilotAfterWeaponTransferFailure(Entity<CombatMechComponent> mech, EntityUid pilot)
+    {
+        Log.Warning($"RX47 ejected {ToPrettyString(pilot)} from {ToPrettyString(mech.Owner)} after weapon transfer failed.");
+
+        if (TryComp(pilot, out BuckleComponent? buckle))
+            _buckle.TryUnbuckle(pilot, pilot, buckle, popup: false);
+
+        UpdateAppearance(mech);
+    }
+}
